Move semi-auto tests to separate package
